Refactor client RPCs from server
