+++ nomad/client_rpc.go
@@ -2,11 +2,17 @@ package nomad
 
 import (
 	"fmt"
+	"io"
+	"net"
+	"sync"
 	"time"
 
 	multierror "github.com/hashicorp/go-multierror"
+	msgpackrpc "github.com/hashicorp/net-rpc-msgpackrpc"
+	"github.com/hashicorp/nomad/helper/pool"
 	"github.com/hashicorp/nomad/nomad/structs"
 	"github.com/hashicorp/yamux"
+	"github.com/ugorji/go/codec"
 )
 
 // nodeConnState is used to track connection information about a Nomad Client.
@@ -116,8 +122,83 @@ func (s *Server) serverWithNodeConn(nodeID string) (*serverParts, error) {
 			return nil, err
 		}
 
-		return nil, ErrNoNodeConn
+		return nil, structs.ErrNoNodeConn
 	}
 
 	return mostRecentServer, nil
+}
+
+// NodeRpc is used to make an RPC call to a node. The method takes the
+// Yamux session for the node and the method to be called.
+func NodeRpc(session *yamux.Session, method string, args, reply interface{}) error {
+	// Open a new session
+	stream, err := session.Open()
+	if err != nil {
+		return err
+	}
+	defer stream.Close()
+
+	// Write the RpcNomad byte to set the mode
+	if _, err := stream.Write([]byte{byte(pool.RpcNomad)}); err != nil {
+		stream.Close()
+		return err
+	}
+
+	// Make the RPC
+	err = msgpackrpc.CallWithCodec(pool.NewClientCodec(stream), method, args, reply)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// NodeStreamingRpc is used to make a streaming RPC call to a node. The method
+// takes the Yamux session for the node and the method to be called. It conducts
+// the initial handshake and returns a connection to be used or an error. It is
+// the callers responsibility to close the connection if there is no error.
+func NodeStreamingRpc(session *yamux.Session, method string) (net.Conn, error) {
+	// Open a new session
+	stream, err := session.Open()
+	if err != nil {
+		return nil, err
+	}
+
+	// Write the RpcNomad byte to set the mode
+	if _, err := stream.Write([]byte{byte(pool.RpcStreaming)}); err != nil {
+		stream.Close()
+		return nil, err
+	}
+
+	// Send the header
+	encoder := codec.NewEncoder(stream, structs.MsgpackHandle)
+	header := structs.StreamingRpcHeader{
+		Method: method,
+	}
+	if err := encoder.Encode(header); err != nil {
+		stream.Close()
+		return nil, err
+	}
+
+	return stream, nil
+}
+
+// Bridge is used to just link two connections together and copy traffic
+func Bridge(a, b io.ReadWriteCloser) error {
+	wg := sync.WaitGroup{}
+	wg.Add(2)
+	go func() {
+		defer wg.Done()
+		io.Copy(a, b)
+		a.Close()
+		b.Close()
+	}()
+	go func() {
+		defer wg.Done()
+		io.Copy(b, a)
+		a.Close()
+		b.Close()
+	}()
+	wg.Wait()
+	return nil
 }
